move core tasktiger logic into its own module
